Ajusta ids nos controllers e services para long

+++ Tickets.Domain/Followup.cs
@@ -3,10 +3,10 @@ using System;
 namespace Tickets.Domain
 {
     public class Followup {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
-        public int? TicketId { get; set; }
+        public long? TicketId { get; set; }
         public virtual Ticket Ticket { get; set; }
         public int UserId { get; set; }
         public virtual User User { get; set; }
