+++ Tickets.WebAPI/Services/ITicketService.cs
@@ -9,8 +9,8 @@ namespace Tickets.WebAPI.Services
     {
          public Task<bool> CreateTicketAsync(Ticket ticket);
          public Task<bool> UpdateTicketAsync(Ticket ticket);
-         public Task<bool> DeleteTicketAsync(int ticketId);
-         public Task<Ticket> GetTicketByIdAsync(int ticketId);
+         public Task<bool> DeleteTicketAsync(long ticketId);
+         public Task<Ticket> GetTicketByIdAsync(long ticketId);
          public Task<ICollection<Ticket>> GetAllTicketsAsync();
     }
 }
